chore: move the about page to the new routes folder
