+++ src/server/router.js
@@ -1,7 +1,7 @@
 import inert from '@hapi/inert'
 
 import { home } from './routes/home/index.js'
-import { about } from './about/index.js'
+import { about } from './routes/about/index.js'
 import { health } from './routes/health/index.js'
 import { serveStaticFiles } from './common/helpers/serve-static-files.js'
